requirements.txt
Flask>=2.0
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Cors
python-dotenv
Werkzeug
passlib